fix(tutorial): align main course entry
